+++ CHANGELOG.md
@@ -15,14 +15,15 @@
         },
       )
     ```
-- add method `tilePropertiesBelow()` and `tilePropertiesListBelow()` in GameComponent to access proprieties of the tile set in Tiled.
+- adds method `tilePropertiesBelow()` and `tilePropertiesListBelow()` in GameComponent to access proprieties of the tile set in Tiled.
+- adds method `void onCollision(GameComponent component)` in `ObjectCollision`. Now you can override this method to listen what Component enter in collision.
 
 # 1.0.3
 
-- Add type `above` in tiled to render above components
+- Adds type `above` in tiled to render above components
 - update `tiledjsonreader`
 - improvements in `Camera`
-- add param `objectsBuilder` in `TiledWorldMap`
+- Adds param `objectsBuilder` in `TiledWorldMap`
 - others improvements
 
 # 1.0.2