+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     ```
 - adds method `tilePropertiesBelow()` and `tilePropertiesListBelow()` in GameComponent to access proprieties of the tile set in Tiled.
 - adds method `void onCollision(GameComponent component, bool active)` in `ObjectCollision`. Now you can override this method to listen what Component enter in collision.
+- improvements in `BonfireGame`
 
 # 1.0.3